refactor: flattened Task folder structure
Flattened folder structure to:
- Domain
  - components
  - Feature 1
  - Feature 2
  - etc.
