+++ src/domains/tasks/components/TaskCategoryBadge.tsx
@@ -1,9 +1,7 @@
-import {
-  TaskStatus,
-  CategoryStatusMap,
-} from "@Tasks/features/ListingsPage/types";
 import { Badge } from "@shopify/polaris";
 import { Status } from "@shopify/polaris/build/ts/latest/src/components/Badge";
+
+import { TaskStatus, CategoryStatusMap } from "@Tasks/ListingsPage/types";
 
 export type TaskCategoryBadgeProps = {
   status: TaskStatus;
